rag/requirements.txt
sentence-transformers>=2.2.2
numpy>=1.21.0
scikit-learn>=1.0.0
faiss-cpu>=1.7.0
huggingface-hub>=0.16.0
torch>=1.9.0
transformers>=4.20.0
